Refactors project into workspace with sub-crates
Refactors the project into a cargo workspace, consisting of the following crates: models, scryfall, provider, archidekt-provider, and rocket-app

This change improves code organization and modularity by separating concerns into independent crates.
