+++ Cargo.toml
@@ -1,40 +1,12 @@
-[package]
-name = "archidekt-live-set-completion"
-version = "0.1.0"
-edition = "2024"
-publish = false
-
-[[bin]]
-name = "archidekt-live-set-completion"
-path = "src/main.rs"
-
-[lib]
-name = "archidekt_live_set_completion_lib"
-path = "src/lib.rs"
-crate-type = ["rlib"]
-
-[build-dependencies]
-reqwest = { version = "0.12", features = ["json","blocking"] }
-chrono = "0.4.42"
+[workspace]
+resolver = '3'
+members = [ "archidekt-provider", "models", "provider","rocket-app", "scryfall"]
+[workspace.dependencies]
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.145"
-lightningcss = "1.0.0-alpha.67"
-
-[dependencies]
-chrono = "0.4.42"
-csv = "1.3.1"
-image = "0.25.8"
-itertools = "0.14.0"
-regex = "1.11.2"
 reqwest = { version = "0.12", features = ["json"] }
-rocket = "0.5.1"
-tokio = { version = "1", features = ["full"] }
-serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.145"
-sha2 = "0.10.9"
-url = "2.5.7"
-urlencoding = "2.1.3"
-
-[dependencies.rocket_dyn_templates]
-version = "0.2.0"
-features = ["handlebars"]
+models = {path = "models"}
+scryfall= {path = "scryfall"}
+provider = {path = "provider"}
+archidekt-provider = {path = "archidekt-provider"}
+tokio = { version = "1", features = ["full"] }
