+++ Rocket.toml
@@ -1,3 +1,9 @@
 [default]
 address = "0.0.0.0"
 port = 10000
+
+[debug]
+template_dir = "rocket-app/templates"
+
+[release]
+template_dir = "rocket-app/templates"
