+++ archidekt-provider/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "archidekt-provider"
+version = "0.1.0"
+edition = "2024"
+
+[lib]
+test=false
+doctest = false
+
+[dependencies]
+models ={ workspace = true}
+reqwest = {workspace = true}
+serde = {workspace = true}
+provider = {workspace = true}
+csv = "1.3.1"
+
+[dev-dependencies]
+tokio = {workspace = true}
