+++ provider/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "provider"
+version = "0.1.0"
+edition = "2024"
+[lib]
+test = true
+doctest = false
+[dependencies]
+models = {workspace = true}
+scryfall = {workspace = true}
+reqwest = {workspace = true}
