+++ rocket-app/templates/index.html.hbs
