+++ rocket-app/templates/layout.html.hbs
